Skills/auth_quickstart/assets/template/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
git+https://github.com/ab0t-com/auth_wrapper.git
